Fix some abs/fabs warnings (possibly abs was also part of f2c and was an fabs in disguise)
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@24340 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ SimulationRuntime/c/optimization/eval_all/EvalG.c
@@ -802,7 +802,7 @@ static inline void debugeJac(OptData * optData, Number* vopt){
            hh *= 0.9;
          else
            break;
-         if(fabs(hh) < 1e-32){
+         if(fabsl(hh) < 1e-32){
            printf("\nWarning: StepSize for FD became very small!\n");
            break;
          }
